Move Fiber app setup function to 'application'
From: 'backend/main.go'
To: 'backend/application/application.go'
Function moved and renamed: 'SetupAppAndSetAuthTo'->'SetupAndSetAuthTo'
