+++ backend/main.go
@@ -2,30 +2,25 @@ package main
 
 import (
 	"log"
-	"strings"
 
-	"github.com/dgrijalva/jwt-go"
-	"github.com/gofiber/fiber/v2"
-
+	"github.com/Slowers-team/Slowers-App/application"
 	"github.com/Slowers-team/Slowers-App/database"
 	"github.com/Slowers-team/Slowers-App/handlers"
 )
 
-var SecretKey []byte
-
 func main() {
-	var databaseURI, port string
-	SecretKey, databaseURI, port = GetEnvironmentVariables()
+	secretKey, databaseURI, port := GetEnvironmentVariables()
 
 	db := database.NewMongoDatabase(databaseURI)
 	if err := db.Connect("Slowers"); err != nil {
 		log.Fatal(err)
 	}
 
-	handlers.SetSecretKey(SecretKey)
+	application.SetSecretKey(secretKey)
+	handlers.SetSecretKey(secretKey)
 	handlers.SetDatabase(db)
 
-	app := SetupAppAndSetAuthTo(true)
+	app := application.SetupAndSetAuthTo(true)
 	app.Static("/", "./client/dist")
 
 	appErr := app.Listen("0.0.0.0:" + port)
@@ -38,48 +33,4 @@ func main() {
 	if dbErr != nil {
 		log.Fatal(dbErr)
 	}
-}
-
-func SetupAppAndSetAuthTo(isAuthOn bool) *fiber.App {
-	app := fiber.New()
-
-	app.Post("/api/register", handlers.CreateUser)
-	app.Post("/api/login", handlers.HandleLogin)
-
-	app.Post("/api/sites", handlers.AddSite)
-	app.Get("/api/sites", handlers.GetRootSites)
-	app.Get("/api/sites/:id", handlers.GetSite)
-	app.Delete("/api/sites/:id", handlers.DeleteSite)
-
-	if (isAuthOn) {
-		app.Use(AuthMiddleware)
-	}
-
-	app.Post("/api/flowers", handlers.AddFlower)
-	app.Get("/api/flowers", handlers.GetFlowers)
-	app.Delete("/api/flowers/:id", handlers.DeleteFlower)
-
-	return app
-}
-
-func AuthMiddleware(c *fiber.Ctx) error {
-	tokenString := c.Get("Authorization")
-
-	if tokenString == "" {
-		return c.Status(fiber.StatusUnauthorized).SendString("Unauthorized")
-	}
-
-	tokenString = strings.TrimPrefix(tokenString, "Bearer ")
-
-	claims := &jwt.StandardClaims{}
-	token, err := jwt.ParseWithClaims(tokenString, claims, func(token *jwt.Token) (interface{}, error) {
-		return SecretKey, nil
-	})
-
-	if err != nil || !token.Valid {
-		return c.Status(fiber.StatusUnauthorized).SendString("Unauthorized")
-	}
-
-	c.Locals("userID", claims.Subject)
-	return c.Next()
 }
